Bug 7247-The BGP configuration is getting configured as "router bgp 0"
Incorporates following changes:
1) Support for 4 byte AS number from the
existing 2 byte.
2) Fix for bgp display commands to work in COMCLI
3) Additionally added debug statement in the code
4) Method onUpdateWithdrawRoute added with nexthop
information. Currently the nexthop information is
not considered when route is withdrawn.

+++ vpnservice/bgpmanager/bgpmanager-impl/src/main/java/org/opendaylight/netvirt/bgpmanager/BgpManager.java
@@ -66,7 +66,7 @@ public class BgpManager implements AutoCloseable, IBgpManager {
     }
 
     public void addNeighbor(String ipAddress, long asNum) throws TException {
-        bcm.addNeighbor(ipAddress, (int) asNum);
+        bcm.addNeighbor(ipAddress, asNum);
     }
 
     public void addEbgpMultihop(String ipAddress, int nhops) throws TException {
@@ -144,7 +144,7 @@ public class BgpManager implements AutoCloseable, IBgpManager {
         bcm.delLogging();
     }
 
-    public void startBgp(int asn, String routerId, int spt, boolean fbit) {
+    public void startBgp(long asn, String routerId, int spt, boolean fbit) {
         bcm.startBgp(asn, routerId, spt, fbit);
     }
